MASSIVE performance improvement for remote hosts.
Previously, the average for `get_mac_address(ip='10.0.0.100')`
was 1.71 seconds. Now, the average is `12.7 miliseconds`, with the
special case of a unpopulated arp table being only slightly higher.

This was brought about by changes in how the arp table is populated.
The original method was to use the host's `ping` command to send an
ICMP packet to the host. This took time, which heavily delayed the
ability to actually get an address. The solution is to instead simply
send a empty UDP packet to a high port.

In hindsight, the `ping` solution made no sense, and was probably the
result of late night coding with little sleep and later on assuming I
was doing it that way for a reason.